+++ lib/llm_cost_tracker/generators/llm_cost_tracker/add_monthly_totals_generator.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+require "rails/generators"
+require "rails/generators/active_record"
+
+module LlmCostTracker
+  module Generators
+    class AddMonthlyTotalsGenerator < Rails::Generators::Base
+      include ActiveRecord::Generators::Migration
+
+      source_root File.expand_path("templates", __dir__)
+
+      desc "Creates a migration to add llm_cost_tracker_monthly_totals"
+
+      def create_migration_file
+        migration_template(
+          "add_monthly_totals_to_llm_cost_tracker.rb.erb",
+          "db/migrate/add_monthly_totals_to_llm_cost_tracker.rb"
+        )
+      end
+
+      private
+
+      def migration_version
+        "[#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}]"
+      end
+    end
+  end
+end